Fix scroll anim crashing on Portfolio because of animated border width
This was not detected on prior versions of RN so it is probably related to the RN upgrade. For now the fix is to simply remove that border as it was invisible anyway (it was using the same color as the background). If we want to reimplement that later on, we will be able to do that by emulating a border at the bottom by using a simple view.